eco_guardian_agent/requirements.txt
google-adk==1.19.0
requests==2.32.5
dotenv==0.9.9
pytrends==4.9.2
streamlit==1.51.0
aiosqlite==0.21.0
jupytext==1.18.1
greenlet==3.2.4